udp-server: Use pthread_cond_wait instead of usleep
Previously in pkt_fwder() if there were no entries in the packet buffer
we'd put the thread to sleep for 100000 usecs. Better is to have the
thread told when there is something in the buffer. For this we can use a
condition variable and the pthread_cond_* functions.

Now what happens is when there is nothing in the queue the threads will
block in pthread_cond_wait(3P) on Linux this is implemented via
futex(2), the main thread will then signal the forwarder threads when
there is work to do.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>