udp-server: Add IPv6 support
You can now specify both IPv4 and IPv6 addresses to bind the server to.
SERVER_IP is renamed to SERVER_IP4 and SERVER_IP6 is introduced. If you
specify them as empty strings, ie "", then the server will bind to all
interfaces for that address family. If you specify the IP as "-" then it
will skip binding for that family.

You can use 'ss -unlp' to see what the server is bound to when it's
running.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>