tools/client: Replace the use of usleep with a nanosleep wrapper
POSIX.1-2001 declares the usleep(3) function obsolete.
POSIX.1-2008 removes its specification.

So we introduce a wrapper around nanosleep(2) (the preferred way to
sleep) called nsleep(). You pass it the number of nanoseconds to sleep
for and it will ensure it sleeps for that long.

We also move the NS_ constants into nsleep.h as they can be used by
anything that wants to use nsleep but doesn't want to have to deal with
the number of nanoseconds in say 250ms, instead you can just pass
NS_MSEC * 250 to nsleep.

Signed-off-by: Andrew Clayton <andrew@dnsandmx.com>